Create fns that store, display and delete tasks
New tasks created with object constructor Task are stored in
an object myTasks. myTasks is an object of objects where task.id
is a key in myTasks object -> myTasks[taskData.id] = taskData.
Each time a new task is added a div is created which is then filled
with task data using Object.entries(taskData).
Also, this div is added into empty property taskData.element
(taskData.element = taskItem) which connects the object with the
corresponding DOM element and enables the getTaskId function which
searches for a match between a clicked node and an object it belongs
to.